requirements.txt
asn1crypto==0.24.0
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
future==0.17.1
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
gsheets==0.3.2
httplib2==0.13.0
idna==2.8
numpy==1.16.4
oauth2client==4.1.3
pandas==0.25.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
pymssql==2.1.4
pyodbc==4.0.26
pyOpenSSL==19.0.0
python-dateutil==2.8.0
python-telegram-bot==11.1.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
rsa==4.0
sentry-sdk==0.3.11
six==1.12.0
SQLAlchemy==1.3.5
uritemplate==3.0.0
urllib3==1.25.3
